+++ .env
@@ -1,2 +1,2 @@
-DATABASE_URL=postgresql+psycopg2://moringa:mikey123@localhost/mazaomarket
+DATABASE_URL=postgresql+psycopg2://@localhost/mazaomarket
 SECRET_KEY=qwertyuiopasdfghjklzxcvbnm
