+++ app/auth/__init__.py
@@ -1,4 +1,4 @@
 from flask import Blueprint
 auth = Blueprint('auth', __name__)
 
-from app.main.views import *
+from .views import *
