+++ app/auth/views.py
@@ -1,9 +1,59 @@
 from app.auth import auth
+from flask import render_template, redirect, url_for, flash, request
+from flask_login import login_user, logout_user, login_required
+from ..models import User
+from .forms import LoginForm
+from ..import db
+from wtforms import StringField, PasswordField, BooleanField, SubmitField
 
-
-@auth.route('/login')
+@auth.route('/login', methods=['GET', 'POST']) # login view function that renders a template file
 def login():
-    pass
+   login_form = LoginForm()
+   if login_form.validate_on_submit():
+       user = User.query.filter_by(email = login_form.email.data).first()
+       if user is not None and user.verify_password(login_form.password.data):
+           login_user(user,login_form.remember.data)
+           return redirect(request.args.get('next') or url_for('main.index'))
+       flash('Invalid username or Password')
+   title = 'watchtower login'
+   return render_template('auth/log in.html', login_form = login_form, title = title)
+
+
+# @auth.route('/logout')
+# @login_required
+# def logout():
+#    logout_user()
+#    # flash('You have been successfully logged out')
+#    return redirect(url_for('main.index'))
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 @auth.route('/signup')
@@ -14,3 +64,4 @@ def signup():
 @auth.route('/logout')
 def logout():
     pass
+
