+++ app/main/__init__.py
@@ -1,5 +1,3 @@
 from flask import Blueprint
-
-main = Blueprint('main', __name__)
-
-from app.main.views import *
+main = Blueprint('main',__name__)
+from . import views
