+++ app/templates/auth/log in.html
@@ -1,22 +1,90 @@
-{% extends 'base.html'%}
-{% import 'bootstrap/wtf.html' as wtf%}
+@auth.route('/login', methods=['GET', 'POST'])
+def login():
+   
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(email=form.email.data).first() 
+        if user and user.verify_password(form.password.data):
+            login_user(user, remember=form.remember.data)
+            next_page = request.args.get('next')
+            return redirect (next_page) if next_page else redirect (url_for('main.index'))
+        else:
+            flash('Login unsuccessful.Please check username and password', 'danger')
+            return redirect(url_for('main.index'))
+
+    return render_template('auth/index.html', title='Login', form=form)
+{% extends "base.html" %}
+
 {% block content %}
-<div class="container">
-   <div class="row" style="margin-top: 50px">
-     <div class= "col-md-2"> </div>
-     <div class="col-md-8">
-     <!--Flash message -->
-     {% for message in get_flashed_messages() %}
-       <div class="alert alert-danger">
-       <button type="button" class="close" data-dismiss="alert">&times;</button>
-       {{ message }}
-       </div>
-     {% endfor %}
-     <!-- Login form -->
-     {{wtf.quick_form(login_form)}}
-     <p> Don't have an account? <a href="{{url_for('auth.register')}}">Sign up here</a></p>
+
+{{super()}}
+
+<div class="container content-section">
+    <form method="POST" action="">
+        {{ form.hidden_tag() }}
+
+        <fieldset class="form-group">
+            <legend class="border-bottom mb-4">Log In</legend>
+
+            <!-- .................................EMAIL field log in............................. -->
+            <div class="form-group">
+                {{ form.email.label(class="form-control-label")}}
+                {% if form.email.errors %}
+                {{ form.email(class="form-control form-control-lg is-invalid")}}
+                <div class="invalid-feedback">
+                    {% for error in form.email.errors%}
+                    <span> {{ error }} </span>
+                    {% endfor %}
+                </div>
+                {% else %}
+                {{ form.email(class="form-control form-control-lg")}}
+                {% endif %}
+            </div>
+
+            <!-- .................................PASWORD FIELD............................. -->
+
+            <div class="form-group">
+                {{ form.password.label(class="form-control-label")}}
+
+                {% if form.password.errors %}
+                {{ form.password(class="form-control form-control-lg is-invalid")}}
+                <div class="invalid-feedback">
+                    {% for error in form.password.errors%}
+                    <span> {{ error }} </span>
+                    {% endfor%}
+                </div>
+                {% else %}
+                {{ form.password(class="form-control form-control-lg")}}
+                {% endif %}
+            </div>
+
+        </fieldset>
+
+        <!-- ...........................Remember a user once they are logged in....................... -->
+        <div class="form-check">
+            {{ form.rememberme(class="form-check-input")}}
+            {{ form.rememberme(class="form-check-label")}}
+        </div>
+
+        <!-- .........................SUBMIT BUTTON............................. -->
+        <div class="form-group">
+            {{ form.submit(class="btn btn-outline-info")}}
+        </div>
+
+    </form>
+
+    <!-- ...................Forgot password.Reset it......................... -->
+    <small class="text-muted">
+        <a href="" class="ml-2">Forgot Password</a>
+    </small>
+
+    <!-- .............................IF No account,redirect to sign up................... -->
+    <div class="border-top">
+        <small class="text-muted">
+            Need an account? <a href="{{ url_for('auth.signup')}}" class="ml-2">Register</a>
+        </small>
     </div>
-    <div clas="col-md-2"> </div>
-   </div>
 </div>
-{% endblock%}
+
+
+{% endblock %}
