+++ requirements.txt
@@ -1,17 +1,22 @@
 alembic==1.2.0
+BareNecessities==0.2.8
 certifi==2019.9.11
 chardet==3.0.4
 Click==7.0
 dominate==2.4.0
 Flask==1.1.1
+Flask-Bootstrap==3.3.7.1
 Flask-Bootstrap4==4.0.2
+Flask-Login==0.4.1
 Flask-Migrate==2.5.2
 Flask-Script==2.0.6
 Flask-SQLAlchemy==2.4.0
 Flask-WTF==0.14.2
+gunicorn==19.9.0
 idna==2.8
 itsdangerous==1.1.0
 Jinja2==2.10.1
+Mail==2.1.0
 Mako==1.1.0
 MarkupSafe==1.1.1
 psycopg2==2.8.3
@@ -24,4 +29,5 @@ SQLAlchemy==1.3.8
 urllib3==1.25.5
 visitor==0.1.3
 Werkzeug==0.16.0
+wtf==0.1
 WTForms==2.2.1
